Added new method for generating tiling to allow more mixed resulting patterns. Leaves gaps for some designs.
palettise_parts is used to determine the closest Lego color of every element in a list of parts and generate the resulting image.
A Part class is defined which includes position, color, and size
The generate_tiling method iterates over corner points of previous parts, selecting the best neighboring part to insert using the best_part method.